cleaned up code architecture
