+++ src/App.css
@@ -82,7 +82,7 @@
   margin-top: 5px;
   margin-bottom: 30px;
   /*padding-bottom: 150px;*/
-  animation: 2s linear 0s 1 fadein;
+  animation: 1s linear 0s 1 fadein;
   clear: left; }
 
 video {
@@ -146,7 +146,6 @@ video {
   left: 0;
   height: 49.25%;
   width: 49.6%;
-  background-position: right;
   background: linear-gradient(0deg, rgba(50, 100, 255, 0.3), rgba(50, 100, 255, 0.1)), url("resources/codeImage.jpg") no-repeat; }
 
 #Design, #skinnyDesign, #shortDesign, #bigDesign {
@@ -154,8 +153,7 @@ video {
   left: 0;
   height: 49.25%;
   width: 49.6%;
-  background-position: right;
-  background: linear-gradient(0deg, rgba(50, 100, 255, 0.3), rgba(50, 100, 255, 0.1)), url("resources/aboutCover.jpg") no-repeat; }
+  background: linear-gradient(0deg, rgba(50, 100, 255, 0.3), rgba(50, 100, 255, 0.1)), url("resources/aboutCover.jpg") no-repeat, left center; }
 
 #bigPhotography, #bigDesign, #bigCoding {
   /*background: linear-gradient(0deg,rgba(50,100,256,0.3),rgba(50,100,256,0.1)), url("resources/computer.jpg") no-repeat;*/
@@ -252,6 +250,7 @@ body {
   font-size: small; }
 
 #social_media_bar {
+  width: 100%;
   text-align: center;
   /*float:left;
     clear:left;*/
@@ -262,62 +261,59 @@ body {
   width: 50px;
   margin: 5px; }
 
-/* Projects Page*/
-.projects_container {
-  width: 100%;
-  padding-bottom: 150px;
-  overflow: hidden; }
-
+/* Projects LIST Page*/
 #coding_projects, #design_projects {
   overflow: auto;
   padding-bottom: 30px;
   border-bottom: 1px solid rgba(0, 0, 0, 0.125); }
 
+.projects {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  align-items: flex-start;
+  width: 95%;
+  overflow: hidden; }
+
 .projects > li {
   list-style: none;
-  /*margin:15px;*/
-  /*padding:25px;*/
-  position: relative;
-  width: 100%;
-  max-width: 475px;
-  float: left;
-  font-size: large;
-  line-height: 30px; }
+  flex-basis: 0;
+  margin: 20px;
+  min-width: 330px;
+  height: 400px;
+  display: flex;
+  flex-direction: column;
+  position: relative; }
 
-.projects > li > a {
-  overflow: auto; }
-
-.projectImage:hover {
-  margin: 0;
-  padding: 25px;
-  padding-bottom: 30px;
-  box-shadow: 10px 10px 50px rgba(40, 40, 126, 0.7);
-  /*transition: .2s ease-out;*/ }
-
-.projectImage, .pageImage {
-  object-fit: contain;
-  margin: 25px;
-  max-width: 425px;
-  max-height: 400px;
-  min-width: 85px;
-  min-height: 80px;
-  width: 100%;
-  height: 100%;
-  margin-bottom: 30px; }
+.projectImageWrap {
+  max-width: 330px;
+  height: 250px;
+  overflow: hidden;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 30px;
+  box-shadow: 5px 10px 40px lightgray; }
 
 .projectImage {
-  box-shadow: 20px 20px 20px rgba(120, 120, 120, 0.3); }
+  max-height: 188px;
+  max-width: 248px;
+  height: auto;
+  width: auto;
+  flex: none; }
 
-.projectText {
-  height: 100px;
-  margin: 0px 25px 25px 25px; }
+.projectImage:hover {
+  max-height: 250px;
+  max-width: 330px;
+  transition: ease-out .5s; }
 
-@media (max-width: 600px) {
+@media (max-width: 780px) {
   .projects {
     overflow: hidden;
     max-width: 100%;
     margin: 0;
-    padding: 0; }
+    padding: 0;
+    justify-content: center; }
   .projectImage {
     max-width: 100%;
     margin: 0 0 15px 0; } }
@@ -328,7 +324,8 @@ body {
 .pageImage {
   padding: 20px;
   /*box-shadow: 20px 20px 50px rgba(120,120,120,.5);*/
-  float: left; }
+  float: left;
+  max-width: 350px; }
 
 .technology {
   float: left;
@@ -372,13 +369,17 @@ h1 {
 .pageGallery {
   clear: left;
   width: 90%;
-  margin: auto; }
+  margin: auto;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center; }
 
 .pageGalleryImage {
   height: auto;
   width: auto;
-  max-width: 30%;
-  padding: 1%;
+  max-width: 300px;
+  max-height: 400px;
+  padding: 25px;
   object-fit: contain; }
 
 .about {
@@ -386,11 +387,9 @@ h1 {
   min-height: 80%; }
 
 .aboutPic {
-  max-width: 300px;
-  float: right;
-  border-radius: 5px;
-  border-bottom-right-radius: 50px;
-  border-top-right-radius: 50px; }
+  max-width: 25%;
+  max-height: 400px;
+  float: right; }
 
 .aboutText {
   max-width: 70%;
