+++ src/Components/Photography.js
@@ -1,0 +1,14 @@
+import React, {Component} from "react";
+
+class Photography extends Component {
+    render() {
+        return (
+            <div>
+                <h1> Photography </h1>
+                <h1> Sorry! Still under construction!  </h1>
+            </div>
+        )
+    }
+}
+
+export {Photography};
